MOBI-489 Showcase Updates avoid NPE on dispose
git-svn-id: http://anonsvn.icesoft.org/repo/icemobile/trunk/icemobile@33111 1de00972-ac04-4911-a46b-06793c8d78d9

+++ samples/mobileshowcase/src/main/java/org/icemobile/samples/mobileshowcase/view/examples/device/camcorder/CamcorderBean.java
@@ -144,9 +144,11 @@ public class CamcorderBean extends ExampleImpl<CamcorderBean> implements
 
     @PreDestroy
     public void disposeResources(){
-        boolean success = camcorderFile.delete();
-        if (!success && logger.isLoggable(Level.FINE)){
-            logger.fine("Could not dispose of media file" + camcorderFile.getAbsolutePath());
+        if( camcorderFile != null ){
+            boolean success = camcorderFile.delete();
+            if (!success && logger.isLoggable(Level.FINE)){
+                logger.fine("Could not dispose of media file" + camcorderFile.getAbsolutePath());
+            }
         }
     }
